+++ src/ignitus-Authentication/ignitus-ProfessorLogin/index.js
@@ -1,6 +1,6 @@
-
 import * as components from './Components';
 
+// eslint-disable-next-line import/no-default-export
 export default {
   components,
 };
