--- src/ignitus-Dashboard/ignitus-StudentDashboard/Components/index.js
@@ -1,2 +1,0 @@
-
-export { default as Dashboard } from './Dashboard';
