--- src/ignitus-Dashboard/ignitus-StudentDashboard/Styles/style.scss
@@ -1,26 +1,0 @@
-// import ignitus token variables
-@import './src/ignitus-Shared/styles/ignitus-tokens';
-
-.marginTop {
-  margin-top: 7rem !important;
-}
-
-.dashboard-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.left-section {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.center-section {
-  grid-area: center-section;
-}
-
-.right-section {
-  grid-area: right-section;
-}
