--- src/ignitus-Dashboard/ignitus-StudentDashboard/index.js
@@ -1,6 +1,0 @@
-
-import * as components from './Components';
-
-export default {
-  components,
-};
