+++ src/ignitus-Dashboard/ignitus-StudentDashboard/index.ts
@@ -1,0 +1,1 @@
+export {};
