+++ src/ignitus-Dashboard/ignitus-StudentDashboard/styles.ts
@@ -1,0 +1,6 @@
+import styled from '@emotion/styled';
+import { BottomSection } from '../../ignitus-Shared/ignitus-DesignSystem/shared';
+
+export const Container = styled(BottomSection)`
+  margin-top: 7rem;
+`;
